+++ CHANGELOG.md
@@ -17,6 +17,7 @@
  - Add CircuitPython variants of the BME280 and SHT31-D Inputs
  - Add support for SHT41x based input devices 
  - Add support for Adafruit's i2c capacitive soil sensor input devices
+ - Add webhook action to emit HTTP requests ([discussion](https://kylegabriel.com/forum/general-discussion/webhook-action/))
  - Add Function Action: MQTT Publish
  - Add ability to set camera stream frames per second
  - Add missing stream resolution option to opencv cameras